requirements.txt
bitcoin-requests
plyvel
cbor2
